+++ src/app/showcase/components/showcase-components.component.html
@@ -87,10 +87,12 @@
     <showcase-title [href]="'context-menu'">Context Menu</showcase-title>
     <showcase-context-menu></showcase-context-menu>
 
-    
+    <showcase-title [href]="'context-menu'">Context Panel</showcase-title>
+    <showcase-context-panel></showcase-context-panel>
+
     <showcase-title [href]="'two-list'">Two-List</showcase-title>
     <showcase-two-list></showcase-two-list>
-   
+
 
     <showcase-title [href]="'loading'">Loading</showcase-title>
     <showcase-loading></showcase-loading>
