+++ src/app/systelab-components/contextmenu/context-menu.component.html
@@ -4,7 +4,7 @@
            [style.color]="fontColor"></i>
     </div>
     <div #dropdownmenu class="dropdown-menu slab-dropdown-menu">
-        <div #dropdown class="slab-dropdown slab-dropdown-fixed">
+        <div #dropdown class="slab-dropdown slab-dropdown-fixed" [style.width]="width">
             <div *ngIf="isOpened">
                 <div #ngcontent (click)="ngContentStopPropagation($event)">
                     <ng-content></ng-content>
