+++ src/app/systelab-components/contextpanel/context-panel.component.html
@@ -1,0 +1,18 @@
+<div #dropdownparent id="{{elementID}}" class="dropdown">
+    <!--  <ng-content select="[header-content]"></ng-content>-->
+
+    <div class="dropdown-toggle slab-context-menu" data-toggle="dropdown" (click)="dotsClicked($event)">
+        <i class="icon-context-menu" [ngClass]="{'d-none': isEmbedded}" [style.font-size]="fontSize"
+           [style.color]="fontColor">Last 30 days</i>
+    </div>
+
+    <div #dropdownmenu class="dropdown-menu slab-dropdown-menu">
+        <div #dropdown class="slab-dropdown slab-dropdown-fixed"><!--[style.width]="width">-->
+            <div *ngIf="isOpened">
+                <div #ngcontent (click)="ngContentStopPropagation($event)">
+                    <ng-content select="[main-content]"></ng-content>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
