+++ src/app/systelab-components/systelab-components.module.ts
@@ -77,6 +77,7 @@ import { ComboboxFavouriteRendererComponent } from './combobox/renderer/combobox
 import { ModulabListBox } from './listbox/listbox.component';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { GridContextMenuComponent } from './grid/contextmenu/grid-context-menu-component';
+import { ContextPanelComponent } from './contextpanel/context-panel.component';
 
 @NgModule({
 	imports:         [
@@ -97,6 +98,7 @@ import { GridContextMenuComponent } from './grid/contextmenu/grid-context-menu-c
 		SliderComponent,
 		SwitchComponent,
 		ContextMenuComponent,
+		ContextPanelComponent,
 		DialogComponent,
 		MessagePopupViewComponent,
 		DataFilterPipe,
@@ -163,6 +165,7 @@ import { GridContextMenuComponent } from './grid/contextmenu/grid-context-menu-c
 		SliderComponent,
 		SwitchComponent,
 		ContextMenuComponent,
+		ContextPanelComponent,
 		TwoListComponent,
 		GridHeaderContextMenuComponent,
 		ColorCellRendererComponent,
